borrow book

+++ src/main/java/com/example/demo/service/book/BookServiceImpl.java
@@ -2,15 +2,18 @@ package com.example.demo.service.book;
 import com.example.demo.model.dto.dtoRequest.BookRequestCreate;
 import com.example.demo.model.dto.dtoResponse.BookResponse;
 import com.example.demo.model.entity.Book;
+import com.example.demo.model.entity.Student;
 import com.example.demo.repository.BookRepository;
+import com.example.demo.repository.StudentRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 import java.util.function.Function;
 
 @Service
@@ -18,6 +21,9 @@ public class BookServiceImpl implements IBookService{
 
     @Autowired
     private BookRepository bookRepository;
+
+    @Autowired
+    private StudentRepository studentRepository;
 
     @Autowired
     private BookResponse bookResponse;
@@ -42,6 +48,9 @@ public class BookServiceImpl implements IBookService{
     @Override
     public List<BookResponse> findAllByNameBookContaining(String name) {
         List<Book> books = bookRepository.findAllByNameBookContaining(name);
+        if (books.isEmpty()){
+            return null;
+        }
         List<BookResponse> bookResponses =new ArrayList<>();
         for ( Book book: books){
             bookResponse = modelMapper.map(book, BookResponse.class);
@@ -53,6 +62,9 @@ public class BookServiceImpl implements IBookService{
     @Override
     public List<BookResponse> findAllByStatusAvailable(String availabe) {
         List<Book> books = bookRepository.findAllByStatus(availabe);
+        if (books.isEmpty()){
+            return null;
+        }
         List<BookResponse> bookResponses =new ArrayList<>();
         for ( Book book: books){
             bookResponse = modelMapper.map(book, BookResponse.class);
@@ -64,6 +76,9 @@ public class BookServiceImpl implements IBookService{
     @Override
     public List<BookResponse> findAllByStatusUnAvailable(String unAvailable) {
         List<Book> books = bookRepository.findAllByStatus(unAvailable);
+        if (books.isEmpty()){
+            return null;
+        }
         List<BookResponse> bookResponses =new ArrayList<>();
         for ( Book book: books){
             bookResponse = modelMapper.map(book, BookResponse.class);
@@ -81,18 +96,58 @@ public class BookServiceImpl implements IBookService{
     }
 
     @Override
-    public BookResponse remove(Long id) {
-        bookEntity = bookRepository.findById(id).get();
-        bookResponse = modelMapper.map(bookEntity, BookResponse.class);
-        bookRepository.deleteById(id);
+    public BookResponse saveBookBorrow(Book book){
+        book.setStartDay(LocalDateTime.now());
+        book.setEndDay(LocalDateTime.now().plusDays(5));
+        book.setStatus("UnAvailable");
+        bookResponse = modelMapper.map(bookRepository.save(book),BookResponse.class);
         return bookResponse;
     }
 
     @Override
+    public BookResponse remove(Long id) {
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.isPresent()){
+            bookResponse = modelMapper.map(bookEntity, BookResponse.class);
+            bookRepository.deleteById(id);
+            return bookResponse;
+        }
+        return null;
+    }
+
+    @Override
     public BookResponse findById(Long id) {
-        bookEntity = bookRepository.findById(id).get();
-        bookResponse = modelMapper.map(bookEntity, BookResponse.class);
-        return bookResponse;
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.isPresent()){
+            bookResponse = modelMapper.map(bookEntity, BookResponse.class);
+            return bookResponse;
+        }
+        return null;
+    }
+
+    @Override
+    public Student borrow(Long idStu, Long idBook){
+        Student student = studentRepository.findById(idStu).get();
+        if (student == null){
+            return null;
+        }
+        if (student.getCheckLegit() == false){
+            return null;
+        }
+        Book book = bookRepository.findById(idBook).get();
+        if (book == null){
+            return null;
+        }
+        if (book.getStatus().equals("UnAvailable")){
+            return null;
+        }
+        book.setStudent(student);
+        book.setStartDay(LocalDateTime.now());
+        book.setEndDay(LocalDateTime.now().plusDays(5));
+        book.setStatus("UnAvailable");
+        bookRepository.save(book);
+        student.getBooks().add(book);
+        return studentRepository.save(student);
     }
 
 }
